clean up schema, make sure foreign keys required

+++ verifact/forum/schema.py
@@ -1,12 +1,12 @@
 from graphene import relay, ObjectType, String, Field, ID, Boolean
 from graphene.relay import Node, ClientIDMutation
+from graphql_relay.node.node import from_global_id
 from graphene_django import DjangoObjectType
 import base64
 from graphql import GraphQLError
 from django.db import IntegrityError
 from django.core.exceptions import ObjectDoesNotExist
 from graphql_jwt.decorators import login_required
-
 from .. import error_strings
 from .models import Question, Answer, Vote
 from verifact.graph.scalars import Url
@@ -48,11 +48,6 @@ class Query(ObjectType):
     def resolve_questions(root, info):
         return Question.objects.all()
 
-    def resolve_answers(root, info):
-        return Answer.objects.all()
-
-    def resolve_votes(root, info):
-        return Vote.objects.all()
 
 class QuestionCreate(ClientIDMutation):
     question = Field(QuestionNode)
@@ -86,7 +81,7 @@ class AnswerCreate(ClientIDMutation):
         text = String()
         citation_url = Url()
         citation_title = String()
-        question_id = ID()
+        question_id = ID(required=True)
 
     @login_required
     def mutate_and_get_payload(
@@ -131,7 +126,7 @@ class VoteUpdate(ClientIDMutation):
 
     class Input:
         credible = Boolean() # if null, remove vote!
-        answer_id = ID()
+        answer_id = ID(required=True)
 
     @login_required
     def mutate_and_get_payload(
@@ -143,31 +138,29 @@ class VoteUpdate(ClientIDMutation):
         credible = kwargs.get('credible',None)
         vote=None
         viewer = info.context.user
-        try: # control flow could have been simpler -> if exists, delete, then create. but chose to do it this way for atomicity.
-            intansid = Node.get_node_from_global_id(
-                info,answer_id,only_type=AnswerNode
-            ).id
-            vote = Vote.objects.get(user=viewer,answer=intansid) # throws ObjectDoesNotExist if none found
+        answer_pk = from_global_id(answer_id)[1]
+        vote_already_exists = True
+        try:
+            vote = Vote.objects.get(user=viewer,answer=answer_pk)
+        except ObjectDoesNotExist as dne:
+            vote_already_exists = False
+
+        if vote_already_exists:
             if credible is None: # if vote null, delete
                 vote.delete()
             else:
                 vote.credible = credible
                 vote.save()
-        except ObjectDoesNotExist as dne:
-            try:
-                if credible is not None:
-                    vote = Vote.objects.create(
-                        user=viewer,
-                        answer=Node.get_node_from_global_id(
-                            info, answer_id, only_type=AnswerNode
-                        ),
-                        credible=credible
-                    )
-            except AssertionError as ae:
-                if str(ae).startswith("Must receive an AnswerNode id."):
-                    raise GraphQLError(error_strings.VOTE_ANSWER_ID_INVALID)
-                else:
-                    raise
+        else:
+            if credible is not None:
+                vote = Vote.objects.create(
+                    user=viewer,
+                    answer=Node.get_node_from_global_id(
+                        info, answer_id, only_type=AnswerNode
+                    ),
+                    credible=credible
+                )
+
 
         return VoteUpdate(vote=vote)
